+++ app/src/main/java/com/babytracker/BabyTrackerApp.kt
@@ -3,6 +3,7 @@ package com.babytracker
 import android.app.Application
 import com.babytracker.core.di.appModule
 import com.babytracker.core.di.databaseModule
+import com.babytracker.core.di.syncModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 
@@ -11,7 +12,7 @@ class BabyTrackerApp : Application() {
         super.onCreate()
         startKoin {
             androidContext(this@BabyTrackerApp)
-            modules(appModule, databaseModule)
+            modules(appModule, databaseModule, syncModule)
         }
     }
 }